Edit the playRound function to take user input and create new 5-round game function
I edited the logic in the playRound function to accept a user's input
and determine whether or not the input beat the computer's random choice,
as well as storing a win/loss tally for the round.

the 5-round game function loops the playRound function and displays
the total scores for the player and the computer at the end of the game.